refactor: 채팅 폴더 구조 및 파일명 개선
- 폴더 구조 개선 (model → hooks, fixtures → mocks)
- 파일명 명확화 (code-link-utils → file-link-parser 등)
- UI 컴포넌트 폴더 구조 개선 (shared → common, message-items → message-types)
- 모든 import 경로 업데이트
