+++ src/features/chat/constants/chat-config.ts
