+++ src/features/chat/constants/file-link-patterns.ts
