+++ src/features/chat/lib/file-link-parser.ts
@@ -1,5 +1,5 @@
-import { FILE_LINK_PATTERN } from '@/features/chat/constants/code-link'
-import type { CodeLink, MessagePart } from '@/features/chat/model/message-types'
+import { FILE_LINK_PATTERN } from '@/features/chat/constants/file-link-patterns'
+import type { CodeLink, MessagePart } from '@/features/chat/types/message-types'
 
 /**
  * 파일명과 라인번호를 표준 형식으로 포맷팅
