+++ src/features/chat/lib/message-helpers.ts
@@ -2,7 +2,7 @@ import {
   MESSAGE_TYPES,
   type Message,
   type SystemMessage,
-} from '@/features/chat/model/message-types'
+} from '@/features/chat/types/message-types'
 
 /**
  * 날짜 헤더 표시 여부를 결정하는 함수
