+++ src/features/chat/ui/common/other-user-avatar.tsx
